refactor(telegram): consolidate bot handlers into named modules
