+++ extensions/telegram/src/bot-handlers.inbound-buffer.ts
@@ -1,4 +1,3 @@
-// Telegram inbound debounce lanes and batch flushing.
 import type { Message } from "grammy/types";
 import { shouldDebounceTextInbound } from "openclaw/plugin-sdk/channel-inbound";
 import {
@@ -6,11 +5,12 @@ import {
   resolveInboundDebounceMs,
 } from "openclaw/plugin-sdk/channel-inbound-debounce";
 import { expectDefined } from "openclaw/plugin-sdk/expect-runtime";
+import { KeyedAsyncQueue } from "openclaw/plugin-sdk/keyed-async-queue";
 import { danger, logVerbose } from "openclaw/plugin-sdk/runtime-env";
-import type { TelegramHandlerMessageRuntime } from "./bot-handlers.message.runtime.js";
+import type { TelegramMessagePipeline } from "./bot-handlers.message-pipeline.js";
+import type { RegisterTelegramHandlerParams } from "./bot-handlers.types.js";
 import type { TelegramMediaRef } from "./bot-message-context.js";
 import type { TelegramAmbientTranscriptWatermark } from "./bot-message-context.types.js";
-import type { RegisterTelegramHandlerParams } from "./bot-native-commands.js";
 import type { TelegramSpooledReplayDeferredParticipant } from "./bot-processing-outcome.js";
 import {
   buildTelegramThreadParams,
@@ -39,10 +39,51 @@ export type TelegramDebounceEntry = {
   spooledReplayParticipant?: TelegramSpooledReplayDeferredParticipant;
 };
 
-export function createTelegramInboundDebounceRuntime(
-  { cfg, bot, runtime }: Pick<RegisterTelegramHandlerParams, "cfg" | "bot" | "runtime">,
-  messageRuntime: TelegramHandlerMessageRuntime,
-) {
+type TextFragmentEntry = {
+  key: string;
+  storeAllowFrom: string[];
+  messages: Array<{ msg: Message; ctx: TelegramContext; receivedAtMs: number }>;
+  promptContextMinTimestampMs?: number;
+  promptContextAmbientWatermark?: TelegramAmbientTranscriptWatermark;
+  dispatchDedupeClaims: TelegramMessageDispatchReplayClaim[];
+  spooledReplayParticipants: TelegramSpooledReplayDeferredParticipant[];
+  timer: ReturnType<typeof setTimeout>;
+};
+
+type TelegramTextFragmentInput = {
+  ctx: TelegramContext;
+  msg: Message;
+  chatId: number;
+  resolvedThreadId?: number;
+  dmThreadId?: number;
+  storeAllowFrom: string[];
+  isAbortControlMessage: boolean;
+  isAuthorizedAbortControlMessage: () => Promise<boolean>;
+  promptContextMinTimestampMs?: number;
+  promptContextAmbientWatermark?: TelegramAmbientTranscriptWatermark;
+  dispatchDedupeClaims: TelegramMessageDispatchReplayClaim[];
+};
+
+interface TelegramInboundBuffers {
+  inboundDebouncer: {
+    enqueue: (entry: TelegramDebounceEntry) => Promise<void>;
+    flushKey: (key: string) => Promise<void>;
+    cancelKey: (key: string) => boolean;
+    drain: () => Promise<void>;
+  };
+  resolveTelegramDebounceEntryMs: (entry: TelegramDebounceEntry) => number;
+  shouldDebounceTelegramEntry: (entry: TelegramDebounceEntry) => boolean;
+  resolveTelegramDebounceLane: (msg: Message) => TelegramDebounceLane;
+  handleTextFragment: (params: TelegramTextFragmentInput) => Promise<boolean>;
+}
+
+export function createTelegramInboundBuffers({
+  params: { cfg, bot, runtime, opts },
+  message,
+}: {
+  params: Pick<RegisterTelegramHandlerParams, "cfg" | "bot" | "runtime" | "opts">;
+  message: TelegramMessagePipeline;
+}): TelegramInboundBuffers {
   const {
     promptContextBoundaryOptions,
     latestPromptContextMinTimestampMs,
@@ -51,12 +92,13 @@ export function createTelegramInboundDebounceRuntime(
     releaseDispatchDedupeClaims,
     buildFailedProcessingResult,
     settleSpooledReplayParticipants,
+    createSpooledReplayParticipantForBufferedWork,
     spooledReplayOptions,
     buildSyntheticTextMessage,
     buildSyntheticContext,
     formatTelegramAmbientTranscriptBody,
     processMessageWithReplyChain,
-  } = messageRuntime;
+  } = message;
   const debounceMs = resolveInboundDebounceMs({ cfg, channel: "telegram" });
   const FORWARD_BURST_DEBOUNCE_MS = 80;
   const resolveTelegramDebounceEntryMs = (entry: TelegramDebounceEntry): number =>
@@ -141,7 +183,6 @@ export function createTelegramInboundDebounceRuntime(
             settleSpooledReplayParticipants(participants, { kind: "skipped" });
             return;
           }
-          // Single entries return above with their original message and structured forward metadata.
           const first = expectDefined(entries.at(0), "multi-entry Telegram debounce batch");
           const syntheticMessage = {
             ...buildSyntheticTextMessage({
@@ -235,10 +276,171 @@ export function createTelegramInboundDebounceRuntime(
     },
   });
 
+  const maxGapMs =
+    typeof opts.testTimings?.textFragmentGapMs === "number" &&
+    Number.isFinite(opts.testTimings.textFragmentGapMs)
+      ? Math.max(10, Math.floor(opts.testTimings.textFragmentGapMs))
+      : 1500;
+  const textBuffer = new Map<string, TextFragmentEntry>();
+  const textQueue = new KeyedAsyncQueue();
+
+  const flushTextFragments = async (entry: TextFragmentEntry) => {
+    try {
+      entry.messages.sort((a, b) => a.msg.message_id - b.msg.message_id);
+      const first = entry.messages[0];
+      const last = entry.messages.at(-1);
+      if (!first || !last) {
+        releaseDispatchDedupeClaims(entry.dispatchDedupeClaims);
+        settleSpooledReplayParticipants(entry.spooledReplayParticipants, { kind: "skipped" });
+        return;
+      }
+      const combinedTextParts = joinTelegramTextParts(
+        entry.messages.map((bufferedMessage) => bufferedMessage.msg),
+        "",
+      );
+      const combinedText = combinedTextParts.text;
+      if (!combinedText.trim()) {
+        releaseDispatchDedupeClaims(entry.dispatchDedupeClaims);
+        settleSpooledReplayParticipants(entry.spooledReplayParticipants, { kind: "skipped" });
+        return;
+      }
+      const syntheticMessage = buildSyntheticTextMessage({
+        base: first.msg,
+        text: combinedText,
+        entities: combinedTextParts.entities,
+        date: last.msg.date ?? first.msg.date,
+      });
+      const result = await processMessageWithReplyChain({
+        ctx: buildSyntheticContext(first.ctx, syntheticMessage),
+        msg: syntheticMessage,
+        allMedia: [],
+        storeAllowFrom: entry.storeAllowFrom,
+        options: {
+          messageIdOverride: String(last.msg.message_id),
+          ambientTranscriptBody: formatTelegramAmbientTranscriptBody(
+            entry.messages.map((bufferedMessage) => bufferedMessage.msg),
+          ),
+          receivedAtMs: first.receivedAtMs,
+          ingressBuffer: "text-fragment",
+          ...promptContextBoundaryOptions(
+            entry.promptContextMinTimestampMs,
+            entry.promptContextAmbientWatermark,
+          ),
+          ...spooledReplayOptions(entry.spooledReplayParticipants),
+        },
+        dispatchDedupeClaims: entry.dispatchDedupeClaims,
+        spooledReplayParticipants: entry.spooledReplayParticipants,
+      });
+      settleSpooledReplayParticipants(entry.spooledReplayParticipants, result);
+    } catch (error) {
+      releaseDispatchDedupeClaims(entry.dispatchDedupeClaims, error);
+      settleSpooledReplayParticipants(
+        entry.spooledReplayParticipants,
+        buildFailedProcessingResult(error),
+      );
+      runtime.error?.(danger(`text fragment handler failed: ${String(error)}`));
+    }
+  };
+  const queueTextFlush = async (entry: TextFragmentEntry) => {
+    await textQueue.enqueue(entry.key, async () => {
+      await flushTextFragments(entry).catch(() => undefined);
+    });
+  };
+  const runTextFlush = async (entry: TextFragmentEntry) => {
+    textBuffer.delete(entry.key);
+    await queueTextFlush(entry);
+  };
+  const scheduleTextFlush = (entry: TextFragmentEntry) => {
+    clearTimeout(entry.timer);
+    entry.timer = setTimeout(() => void runTextFlush(entry), maxGapMs);
+  };
+  const handleTextFragment = async (params: TelegramTextFragmentInput): Promise<boolean> => {
+    const text = typeof params.msg.text === "string" ? params.msg.text : undefined;
+    const isCommandLike = (text ?? "").trim().startsWith("/");
+    const senderId = params.msg.from?.id != null ? String(params.msg.from.id) : "unknown";
+    const threadId = params.resolvedThreadId ?? params.dmThreadId;
+    const key = `text:${params.chatId}:${threadId ?? "main"}:${senderId}`;
+    if (text && !isCommandLike && !params.isAbortControlMessage) {
+      const nowMs = Date.now();
+      const existing = textBuffer.get(key);
+      if (existing) {
+        const last = existing.messages.at(-1);
+        const idGap = last ? params.msg.message_id - last.msg.message_id : Infinity;
+        const timeGapMs = nowMs - (last?.receivedAtMs ?? nowMs);
+        const canAppend = idGap > 0 && idGap <= 1 && timeGapMs >= 0 && timeGapMs <= maxGapMs;
+        const nextTotalChars =
+          existing.messages.reduce(
+            (sum, bufferedMessage) => sum + (bufferedMessage.msg.text?.length ?? 0),
+            0,
+          ) + text.length;
+        if (canAppend && existing.messages.length < 12 && nextTotalChars <= 50_000) {
+          const participant = createSpooledReplayParticipantForBufferedWork(
+            `text-fragment:${key}:${params.msg.message_id}`,
+          );
+          if (participant) {
+            existing.spooledReplayParticipants.push(participant);
+          }
+          existing.messages.push({ msg: params.msg, ctx: params.ctx, receivedAtMs: nowMs });
+          existing.promptContextMinTimestampMs = latestPromptContextMinTimestampMs(
+            existing.promptContextMinTimestampMs,
+            params.promptContextMinTimestampMs,
+          );
+          existing.promptContextAmbientWatermark = latestPromptContextAmbientWatermark(
+            existing.promptContextAmbientWatermark,
+            params.promptContextAmbientWatermark,
+          );
+          existing.dispatchDedupeClaims = mergeDispatchDedupeClaims(
+            existing.dispatchDedupeClaims,
+            params.dispatchDedupeClaims,
+          );
+          scheduleTextFlush(existing);
+          return true;
+        }
+        clearTimeout(existing.timer);
+        textBuffer.delete(key);
+        await queueTextFlush(existing);
+      }
+      if (text.length >= 4000) {
+        const participant = createSpooledReplayParticipantForBufferedWork(
+          `text-fragment:${key}:${params.msg.message_id}`,
+        );
+        const entry: TextFragmentEntry = {
+          key,
+          storeAllowFrom: params.storeAllowFrom,
+          messages: [{ msg: params.msg, ctx: params.ctx, receivedAtMs: nowMs }],
+          dispatchDedupeClaims: params.dispatchDedupeClaims,
+          spooledReplayParticipants: participant ? [participant] : [],
+          ...promptContextBoundaryOptions(
+            params.promptContextMinTimestampMs,
+            params.promptContextAmbientWatermark,
+          ),
+          timer: setTimeout(() => {}, maxGapMs),
+        };
+        textBuffer.set(key, entry);
+        scheduleTextFlush(entry);
+        return true;
+      }
+    } else if (
+      text &&
+      params.isAbortControlMessage &&
+      (await params.isAuthorizedAbortControlMessage())
+    ) {
+      const existing = textBuffer.get(key);
+      if (existing) {
+        clearTimeout(existing.timer);
+        textBuffer.delete(key);
+        releaseDispatchDedupeClaims(existing.dispatchDedupeClaims);
+        settleSpooledReplayParticipants(existing.spooledReplayParticipants, { kind: "skipped" });
+      }
+    }
+    return false;
+  };
+
   return {
     inboundDebouncer,
     resolveTelegramDebounceEntryMs,
     shouldDebounceTelegramEntry,
     resolveTelegramDebounceLane,
+    handleTextFragment,
   };
 }
