+++ extensions/telegram/src/bot-handlers.reaction.runtime.test.ts
@@ -3,9 +3,10 @@ import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
 import { getChildLogger } from "openclaw/plugin-sdk/runtime-env";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { defaultTelegramBotDeps } from "./bot-deps.js";
-import { createTelegramHandlerAuthorizationRuntime } from "./bot-handlers.authorization.runtime.js";
-import { registerTelegramReactionHandler } from "./bot-handlers.reaction.runtime.js";
-import type { RegisterTelegramHandlerParams } from "./bot-native-commands.js";
+import { createTelegramEventBindings } from "./bot-handlers.event-bindings.js";
+import { createTelegramHandlerAuthorization } from "./bot-handlers.inbound-authorization.js";
+import { createTelegramMessagePipeline } from "./bot-handlers.message-pipeline.js";
+import type { RegisterTelegramHandlerParams } from "./bot-handlers.types.js";
 import type { TelegramThreadSpec } from "./bot/helpers.js";
 
 const FIRE_EMOJI = "\u{1F525}";
@@ -106,11 +107,15 @@ function registerHandler(cfg: OpenClawConfig): ReactionHandler {
     },
   };
 
-  registerTelegramReactionHandler(
+  createTelegramEventBindings({
     params,
-    { resolveCachedMessageThreadSpec },
-    createTelegramHandlerAuthorizationRuntime(params),
-  );
+    message: {
+      ...createTelegramMessagePipeline(params),
+      resolveCachedMessageThreadSpec,
+    },
+    authorization: createTelegramHandlerAuthorization(params),
+    registerMessages: () => {},
+  }).registerReaction();
   const handler = handlers.get("message_reaction");
   if (!handler) {
     throw new Error("expected message_reaction handler");
